[WIP] -Added new files -

+++ CiceSports/Managers/TableView/TablePresenterProtocol.swift
@@ -14,13 +14,13 @@ protocol TablePresenterProtocol: class {
     func numberOfCell(_ tableType: TableType, section: Int) -> Int
     func object(_ tableType: TableType, indexPath: IndexPath) -> Any
     func didTapRow(_ tableType: TableType, indexPath: IndexPath)
-    func heightForRow(_ tableType: TableType, indexPath: IndexPath) -> Int
+    func heightForRow(_ tableType: TableType, indexPath: IndexPath) -> CGFloat
     
 }
 
 extension TablePresenterProtocol {
     func didTapRow(_ tableType: TableType, indexPath: IndexPath) {}
-    func heightForRow(_ tableType: TableType, indexPath: IndexPath) -> Int { return 44 }
+    func heightForRow(_ tableType: TableType, indexPath: IndexPath) -> CGFloat { return 44 }
 }
 
 protocol TablePresenterDelegate: class {
